requirements.txt.txt
alembic==1.12.0
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2022.1
distlib==0.3.7
Django==4.2.5
django-crispy-forms==2.0
dnspython==2.4.2
email-validator==2.0.0.post2
filelock==3.12.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
Pillow==10.0.0
pipenv==2023.9.1
platformdirs==3.10.0
pluggy==1.3.0
pytest==7.4.2
pytest-flask==1.2.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
virtualenv==20.24.4
Werkzeug==2.3.7
WTForms==3.0.1